requirements.txt
appdirs==1.4.4
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.3
colorama==0.4.4
cssselect==1.1.0
fake-useragent==0.1.11
idna==3.2
lxml==4.6.3
parse==1.19.0
pyee==8.1.0
pyppeteer==0.2.5
pyquery==1.4.3
requests==2.26.0
requests-html==0.10.0
six==1.16.0
soupsieve==2.2.1
tqdm==4.61.2
urllib3==1.25.11
w3lib==1.22.0
websockets==8.1
python-dotenv~=0.19.0